+++ .eslintrc.js
@@ -7,6 +7,8 @@ module.exports = {
     "eslint:recommended",
     "plugin:react/recommended",
     "plugin:@typescript-eslint/recommended",
+    "plugin:import/recommended",
+    "plugin:import/typescript",
     "prettier",
   ],
   overrides: [],
@@ -20,10 +22,11 @@ module.exports = {
       version: "detect",
     },
   },
-  plugins: ["react", "@typescript-eslint"],
+  plugins: ["react", "@typescript-eslint", "import"],
   rules: {
     "no-console": "error",
     // Necessary to work with react-three-fiber. See: https://github.com/jsx-eslint/eslint-plugin-react/issues/3423
     "react/no-unknown-property": "off",
+    "import/order": ["error", { alphabetize: { order: "asc" } }],
   },
 };
