+++ package.json
@@ -44,6 +44,7 @@
     "@typescript-eslint/eslint-plugin": "^5.59.5",
     "eslint": "^8.40.0",
     "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-react": "^7.32.2",
     "prettier": "2.8.8"
   }
