+++ src/App.tsx
@@ -1,15 +1,19 @@
 import React, { FC, useEffect, useState } from "react";
+import { DebugContext } from "./domain/Debug";
 import {
   DimensionsContext,
   blockSpace,
   headerSize,
   marginSize,
 } from "./domain/Dimensions";
-import { Header } from "./ui/components/header/Header";
 import { Canvas } from "./ui/components/canvas/Canvas";
+import { Header } from "./ui/components/header/Header";
 import "./App.css";
 
 export const App: FC = () => {
+  const [metaKeyDown, setMetaKeyDown] = useState(false);
+  const [alternateKeyDown, setAlternateKeyDown] = useState(false);
+  const [debugMode, setDebugMode] = useState(false);
   const [height, setHeight] = useState(
     // Canvas height is the window height minus the header height minus two times
     // the margin size for the page margins and one more for the canvas top margin.
@@ -23,7 +27,6 @@ export const App: FC = () => {
   // The sceneHeight sets the whole height of the scene in its own coordinate unit
   const [sceneHeight, setSceneHeight] = useState(0);
   const aspect = width / height;
-
   useEffect(() => {
     const handleResize = () => {
       setHeight(window.innerHeight - headerSize - 3 * marginSize);
@@ -32,6 +35,40 @@ export const App: FC = () => {
     window.addEventListener("resize", handleResize);
     return () => window.removeEventListener("resize", handleResize);
   }, []);
+
+  useEffect(() => {
+    const keyDownEventListener = (e: KeyboardEvent) => {
+      if (e.key === "Control") {
+        setMetaKeyDown(true);
+      }
+      if (e.key === "d") {
+        setAlternateKeyDown(true);
+      }
+      return false;
+    };
+    document.addEventListener("keydown", keyDownEventListener);
+
+    const keyUpEventListener = (e: KeyboardEvent) => {
+      if (e.key === "Control") {
+        setMetaKeyDown(false);
+      }
+      if (e.key === "d") {
+        setAlternateKeyDown(false);
+      }
+    };
+    document.addEventListener("keyup", keyUpEventListener);
+
+    return () => {
+      document.removeEventListener("keydown", keyDownEventListener);
+      document.removeEventListener("keyup", keyUpEventListener);
+    };
+  }, []);
+
+  useEffect(() => {
+    if (metaKeyDown && alternateKeyDown) {
+      setDebugMode(!debugMode);
+    }
+  }, [metaKeyDown, alternateKeyDown]);
 
   return (
     <main className="App">
@@ -47,8 +84,10 @@ export const App: FC = () => {
           setSceneWidth,
         }}
       >
-        <Header />
-        <Canvas />
+        <DebugContext.Provider value={{ debug: debugMode }}>
+          <Header />
+          <Canvas />
+        </DebugContext.Provider>
       </DimensionsContext.Provider>
     </main>
   );
